+++ HISTORY.md
@@ -12,6 +12,7 @@
     - `itransfer-variant-calling`
 - Adding `mypy` checks to CI.
 - Adding `--dry-run` and `--show-diff` arguments to `cubi-sak snappy pull-sheet`.
+- Adding `cubi-sak snake check` command.
 
 ## v0.1.0